Changed namespace and added overloads for HTTP aliases.
